requirements.txt
google-generativeai==0.3.2
PyGithub==1.59.0
python-dotenv==1.0.0
requests==2.28.2